+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `--parent` flag for recording/capturing the parent window ([#14](https://github.com/orhun/menyoki/issues/10))
+
 ### Changed
 - Use architecture compatible types for Xlib calls